Refactor: Move AuthRepository to repository package
The AuthRepository class has been moved from the `model` package to a new `repository` package. This change aims to improve the project structure by separating data access logic (repositories) from data models.

The ViewModels (LoginViewModel and SignupViewModel) have been updated to reflect this change in the import path for AuthRepository.
