+++ app/src/main/java/com/example/noctaleapp/viewmodel/LoginViewModel.kt
@@ -3,7 +3,7 @@ package com.example.noctaleapp.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.noctaleapp.model.AuthRepository
+import com.example.noctaleapp.repository.AuthRepository
 
 class LoginViewModel : ViewModel() {
     private val repository = AuthRepository()
